+++ crates/tui-realm-stdlib/examples/canvas.rs
@@ -35,6 +35,7 @@ use std::time::{Duration, Instant};
 
 use tui_realm_stdlib::components::{label, Canvas, CanvasPropsBuilder};
 use tuirealm::props::borders::{BorderType, Borders};
+use tuirealm::props::Alignment;
 use tuirealm::tui::widgets::canvas::MapResolution;
 use tuirealm::{Msg, PropsBuilder, Update, View};
 // tui
@@ -92,7 +93,7 @@ fn main() {
                 .hidden()
                 .visible()
                 .with_background(Color::Reset)
-                .with_title(String::from("playing risiko"))
+                .with_title(String::from("playing risiko"), Alignment::Center)
                 .with_borders(Borders::ALL, BorderType::Rounded, Color::LightYellow)
                 .with_x_bounds((-180.0, 180.0))
                 .with_y_bounds((-90.0, 90.0))
@@ -171,7 +172,7 @@ impl Update for Model {
         match ref_msg {
             None => None, // Exit after None
             Some(msg) => match msg {
-                (_, &MSG_KEY_ESC) => {
+                (_, key) if key == &MSG_KEY_ESC => {
                     // Quit on esc
                     self.quit();
                     None